+++ Microsoft.Azure.Cosmos/src/Diagnostics/CosmosDiagnosticsContextCore.cs
@@ -4,7 +4,6 @@
 namespace Microsoft.Azure.Cosmos
 {
     using System;
-    using System.Collections;
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Diagnostics;
@@ -26,7 +25,7 @@ namespace Microsoft.Azure.Cosmos
 
         private static readonly string DefaultUserAgentString;
 
-        private bool IsDefaultUserAgent = true;
+        private readonly bool IsDefaultUserAgent = true;
 
         private bool isOverallScopeSet = false;
 
@@ -93,14 +92,29 @@ namespace Microsoft.Azure.Cosmos
                 throw new ArgumentNullException(nameof(pointOperationStatistics));
             }
 
-            this.TotalRequestCount++;
-            int statusCode = (int)pointOperationStatistics.StatusCode;
-            if (statusCode < 200 || statusCode > 299)
-            {
-                this.FailedRequestCount++;
-            }
+            this.AddRequestCount((int)pointOperationStatistics.StatusCode);
 
             this.ContextList.Add(pointOperationStatistics);
+        }
+
+        internal override void AddDiagnosticsInternal(StoreResponseStatistics storeResponseStatistics)
+        {
+            if (storeResponseStatistics.StoreResult != null)
+            {
+                this.AddRequestCount((int)storeResponseStatistics.StoreResult.StatusCode);
+            }
+
+            this.ContextList.Add(storeResponseStatistics);
+        }
+
+        internal override void AddDiagnosticsInternal(AddressResolutionStatistics addressResolutionStatistics)
+        {
+            this.ContextList.Add(addressResolutionStatistics);
+        }
+
+        internal override void AddDiagnosticsInternal(CosmosClientSideRequestStatistics clientSideRequestStatistics)
+        {
+            this.ContextList.Add(clientSideRequestStatistics);
         }
 
         internal override void AddDiagnosticsInternal(QueryPageDiagnostics queryPageDiagnostics)
@@ -143,6 +157,15 @@ namespace Microsoft.Azure.Cosmos
         public override IEnumerator<CosmosDiagnosticsInternal> GetEnumerator()
         {
             return this.ContextList.GetEnumerator();
+        }
+
+        private void AddRequestCount(int statusCode)
+        {
+            this.TotalRequestCount++;
+            if (statusCode < 200 || statusCode > 299)
+            {
+                this.FailedRequestCount++;
+            }
         }
 
         private void AddSummaryInfo(CosmosDiagnosticsContext newContext)