+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosQueryUnitTests.cs
@@ -347,13 +347,13 @@ namespace Microsoft.Azure.Cosmos.Tests
                     Guid.NewGuid().ToString(),
                     System.Net.HttpStatusCode.Unauthorized,
                     subStatusCode: SubStatusCodes.PartitionKeyMismatch,
+                    responseTimeUtc: DateTime.UtcNow,
                     requestCharge: 4,
                     errorMessage: null,
                     method: HttpMethod.Post,
                     requestUri: new Uri("http://localhost.com"),
                     requestSessionToken: null,
-                    responseSessionToken: null,
-                    clientSideRequestStatistics: null));
+                    responseSessionToken: null));
             IReadOnlyCollection<QueryPageDiagnostics> diagnostics = new List<QueryPageDiagnostics>()
             {
                 new QueryPageDiagnostics(